% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{AnnotateRun-class}
\alias{AnnotateRun-class}
\title{Annotate-hits result.}
\description{
Annotate-hits result.
}
\section{Slots}{

\describe{
\item{\code{n_input_hits}}{Input hit count.}

\item{\code{n_genes}}{Unique annotated genes.}

\item{\code{crop}}{Crop name (when used).}

\item{\code{assembly}}{Assembly accession (when used).}

\item{\code{window_bp}}{Annotation window in base pairs.}

\item{\code{hits}}{Tibble of annotated hits.}

\item{\code{genes}}{Tibble of unique genes.}
}}

