% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{ClumpRun-class}
\alias{ClumpRun-class}
\title{LD-clumping result.}
\description{
LD-clumping result.
}
\section{Slots}{

\describe{
\item{\code{n_input_variants}}{Variants in input sumstats.}

\item{\code{n_clumps}}{Clumps formed.}

\item{\code{n_index_variants}}{Index-variant count.}

\item{\code{p_threshold}}{P-value threshold used.}

\item{\code{r2_threshold}}{r^2 threshold used.}

\item{\code{clumps}}{Tibble of clumps.}
}}

