% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{ConvertRun-class}
\alias{ConvertRun-class}
\title{Convert result.}
\description{
Convert result.
}
\section{Slots}{

\describe{
\item{\code{input_format}}{Source format.}

\item{\code{output_format}}{Destination format.}

\item{\code{n_samples}}{Sample count.}

\item{\code{n_variants}}{Variant count.}
}}

