% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compact}
\alias{.compact}
\title{Strip NULL-valued entries from a list (for arg passthrough).}
\usage{
.compact(x)
}
\description{
Strip NULL-valued entries from a list (for arg passthrough).
}
\keyword{internal}
