% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_runtime.R
\name{.managed_venv_name}
\alias{.managed_venv_name}
\title{Locate the managed Python environment.}
\usage{
.managed_venv_name()
}
\description{
Returns the name of the rTorchGenomics-managed virtualenv.
}
\keyword{internal}
