% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{data}
\name{IClassResult}
\alias{IClassResult}
\title{Constructor for the IClassResult S4 class.}
\format{
An object of class \code{list} of length 1.
}
\usage{
IClassResult
}
\description{
Wrapper list exposing \verb{$IClassResult$new_from_dict(d)} for building
an instance from a list of py_to_r'd attributes from the Python
\code{IClassResult} dataclass.
}
\keyword{internal}
