% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{ImputeRun-class}
\alias{ImputeRun-class}
\title{Impute result.}
\description{
Impute result.
}
\section{Slots}{

\describe{
\item{\code{method}}{Imputation method.}

\item{\code{n_samples}}{Sample count.}

\item{\code{n_variants}}{Variant count.}

\item{\code{n_imputed}}{Imputed-genotype count.}

\item{\code{imputation_quality}}{Aggregate quality metric (e.g., r^2).}
}}

