% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{LDBlocksRun-class}
\alias{LDBlocksRun-class}
\title{LD-blocks result.}
\description{
LD-blocks result.
}
\section{Slots}{

\describe{
\item{\code{method}}{LD block detection method.}

\item{\code{n_blocks}}{Detected block count.}

\item{\code{n_variants_in_blocks}}{Variants assigned to a block.}

\item{\code{median_block_size_bp}}{Median block width in base pairs.}

\item{\code{median_block_n_snps}}{Median variants per block.}

\item{\code{blocks}}{Tibble of detected blocks (CHR / start / end / n_snps).}
}}

