% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{MRMegaRun-class}
\alias{MRMegaRun-class}
\title{Multi-ancestry MR meta-analysis (MR-MEGA) result.}
\description{
Multi-ancestry MR meta-analysis (MR-MEGA) result.
}
\section{Slots}{

\describe{
\item{\code{method}}{Meta-analysis method (always "mr_mega").}

\item{\code{n_axes}}{Number of MEGA axes of genetic variation used.}

\item{\code{n_populations}}{Number of input populations/studies.}

\item{\code{n_snps}}{Number of SNPs in the meta-analysis.}

\item{\code{min_p_meta}}{Smallest meta-analysis p-value observed.}

\item{\code{results}}{Tibble of per-SNP meta-analysis statistics.}

\item{\code{raw}}{Full result dict, for fields not otherwise exposed.}
}}

