% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\name{output_files}
\alias{output_files}
\alias{output_files,_BaseRun-method}
\title{Extract output file paths from a result object.}
\usage{
output_files(x)

\S4method{output_files}{_BaseRun}(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
A named list of paths.
}
\description{
Extract output file paths from a result object.
}
