% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{data}
\name{PowerRun}
\alias{PowerRun}
\title{Constructor for the PowerRun S4 class.}
\format{
An object of class \code{list} of length 1.
}
\usage{
PowerRun
}
\description{
Wrapper list exposing \verb{$PowerRun$new_from_dict(d)} for building an
instance from a JSON-shaped list (the dict returned by
\code{py_to_r(to_dict())}).
}
\keyword{internal}
