% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\docType{class}
\name{ScanRun-class}
\alias{ScanRun-class}
\title{Scan-run result (lmm_scan, glm_scan).}
\description{
Scan-run result (lmm_scan, glm_scan).
}
\section{Slots}{

\describe{
\item{\code{model}}{Model name (e.g., SingleTraitLMM).}

\item{\code{test}}{Test statistic name (wald, score, lrt).}

\item{\code{correction}}{Multiple-testing correction (bh, by, bonferroni, ...).}

\item{\code{n_variants}}{Variant count tested.}

\item{\code{n_significant}}{Variant count exceeding significance threshold.}

\item{\code{significance_threshold}}{P-value threshold.}

\item{\code{lambda_gc}}{Genomic-control inflation factor.}

\item{\code{sigma2_g}}{Additive genetic variance estimate.}

\item{\code{sigma2_e}}{Residual variance estimate.}

\item{\code{h2}}{Narrow-sense heritability estimate.}

\item{\code{n_samples}}{Sample count used.}

\item{\code{top_hits}}{Tibble of top variants by p-value.}
}}

