% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tg_mr_mega}
\alias{tg_mr_mega}
\title{Multi-ancestry MR meta-analysis (MR-MEGA).}
\usage{
tg_mr_mega(sumstats, output = NULL, n_axes = 4L, random_effects = FALSE)
}
\arguments{
\item{sumstats}{Character vector of >= 3 per-population MR sumstats
paths.}

\item{output}{Optional path to write the results TSV.}

\item{n_axes}{Number of MEGA axes of genetic variation to regress out.}

\item{random_effects}{If \code{TRUE}, use a random-effects meta-analysis
term in addition to the MEGA-axis regression.}
}
\value{
An \code{MRMegaRun} S4 object.
}
\description{
Multi-ancestry MR meta-analysis (MR-MEGA).
}
\examples{
\dontrun{
r <- tg_mr_mega(c("pop1.tsv", "pop2.tsv", "pop3.tsv"), n_axes = 2)
}
}
