% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tg_qq}
\alias{tg_qq}
\title{Q-Q plot of GWAS p-values.}
\usage{
tg_qq(x, title = NULL, p_col = "P")
}
\arguments{
\item{x}{A \code{ScanRun} or compatible object.}

\item{title}{Plot title.}

\item{p_col}{Column name for p-values.}
}
\value{
A \code{ggplot} object.
}
\description{
Q-Q plot of GWAS p-values.
}
\examples{
\dontrun{
tg_qq(tg_lmm_scan("data.bed", "pheno.tsv"))
}
}
