% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\name{to_dataframe}
\alias{to_dataframe}
\alias{to_dataframe,LGEBVResult-method}
\alias{to_dataframe,IClassResult-method}
\title{Convert a result object to a per-row tibble / data.frame.}
\usage{
to_dataframe(x)

\S4method{to_dataframe}{LGEBVResult}(x)

\S4method{to_dataframe}{IClassResult}(x)
}
\arguments{
\item{x}{A result object (e.g., \code{LGEBVResult}, \code{IClassResult}).}
}
\value{
A tibble.
}
\description{
Convert a result object to a per-row tibble / data.frame.
}
