% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_classes.R
\name{top_hits}
\alias{top_hits}
\alias{top_hits,ScanRun-method}
\alias{top_hits,GwasResult-method}
\title{Extract the top-hits tibble from a result object.}
\usage{
top_hits(x)

\S4method{top_hits}{ScanRun}(x)

\S4method{top_hits}{GwasResult}(x)
}
\arguments{
\item{x}{A ScanRun or GwasResult.}
}
\value{
A tibble.
}
\description{
Extract the top-hits tibble from a result object.
}
